updated chat feature v1 to mvvm design
