+++ SniffServe/Models/ChatMessage.swift
@@ -6,3 +6,10 @@
 //
 
 import Foundation
+
+struct ChatMessage: Codable, Identifiable {
+    let id = UUID()
+    let role: String
+    let content: String
+}
+
