+++ SniffServe/Models/OpenAIResponse.swift
@@ -6,3 +6,11 @@
 //
 
 import Foundation
+
+struct OpenAIResponse: Codable {
+    struct Choice: Codable {
+        let message: ChatMessage
+    }
+    let choices: [Choice]
+}
+
