+++ SniffServe/Views/ChatView.swift
@@ -7,33 +7,11 @@
 
 import SwiftUI
 
-// Define the data structures for handling the API interactions
-struct Message: Codable, Identifiable {
-    let id = UUID()
-    let role: String
-    let content: String
-}
-
-struct ChatRequest: Codable {
-    let model: String
-    let messages: [Message]
-}
-
-struct ChatResponse: Codable {
-    struct Choice: Codable {
-        let message: Message
-    }
-    let choices: [Choice]
-}
-
-// ChatView with integrated chat functionality
 struct ChatView: View {
-    @State private var userInput: String = ""
-    @State private var messages: [Message] = []
+    @StateObject private var viewModel = OpenAIViewModel()
 
     var body: some View {
         VStack {
-            // Chat header, can replace or modify according to the layout needs
             HStack {
                 Button(action: {}) {
                     Image(systemName: "chevron.backward")
@@ -56,7 +34,7 @@ struct ChatView: View {
 
             ScrollView {
                 LazyVStack(spacing: 10) {
-                    ForEach(messages) { message in
+                    ForEach(viewModel.messages) { message in
                         MessageView(message: message)
                     }
                 }
@@ -65,12 +43,12 @@ struct ChatView: View {
 
             // Input area for sending messages
             HStack {
-                TextField("Type your message here...", text: $userInput)
+                TextField("Type your message here...", text: $viewModel.userInput)
                     .textFieldStyle(RoundedBorderTextFieldStyle())
                     .padding()
 
                 Button("Send") {
-                    sendMessage()
+                    viewModel.sendMessage()
                 }
                 .padding()
                 .background(Color.blue)
@@ -80,26 +58,11 @@ struct ChatView: View {
             .padding()
         }
     }
-
-    func sendMessage() {
-        let apiKey = ProcessInfo.processInfo.environment["OPENAI_KEY"] ?? "defaultKey"  // Use the actual API key here
-        sendChatMessage(apiKey: apiKey, message: userInput) { reply in
-            DispatchQueue.main.async {
-                if let reply = reply {
-                    let userMessage = Message(role: "user", content: userInput)
-                    let aiMessage = Message(role: "ai", content: reply)
-                    messages.append(userMessage)
-                    messages.append(aiMessage)
-                    userInput = ""
-                }
-            }
-        }
-    }
 }
 
 // Custom view for displaying messages
 struct MessageView: View {
-    var message: Message
+    var message: ChatMessage
 
     var body: some View {
         HStack {
@@ -121,45 +84,6 @@ struct MessageView: View {
         }
         .padding(.horizontal)
     }
-}
-
-// Function to handle sending messages to OpenAI API
-func sendChatMessage(apiKey: String, message: String, completion: @escaping @Sendable (String?) -> Void) {
-    let url = URL(string: "https://api.openai.com/v1/chat/completions")!
-    let userMessage = Message(role: "user", content: message)
-    let requestBody = ChatRequest(model: "gpt-3.5-turbo", messages: [userMessage])
-
-    guard let jsonData = try? JSONEncoder().encode(requestBody) else {
-        completion(nil)
-        return
-    }
-
-    var request = URLRequest(url: url)
-    request.httpMethod = "POST"
-    request.httpBody = jsonData
-    request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-    request.addValue("Bearer \(apiKey)", forHTTPHeaderField: "Authorization")
-
-    let task = URLSession.shared.dataTask(with: request) { data, response, error in
-        if let error = error {
-            print("Error making request: \(error)")
-            completion(nil)
-            return
-        }
-
-        guard let data = data else {
-            completion(nil)
-            return
-        }
-
-        if let decodedResponse = try? JSONDecoder().decode(ChatResponse.self, from: data) {
-            let reply = decodedResponse.choices.first?.message.content
-            completion(reply)
-        } else {
-            completion(nil)
-        }
-    }
-    task.resume()
 }
 
 // Preview provider for SwiftUI preview
